Move project to the new debhelper dependency
The debian/compat file is not recommended to configure debhelper as a
build dependency. Therefore, the aformenetioned package has been deleted
and replaced with the debhelper-compat (= 13) dependency stated in the
debian/control file.

The recommended debhelper level for Ubuntu Jammy (22.04 LTS) is 13. This
required a version bump from the older level, 10.

For more information, please check:
* https://nthykier.wordpress.com/2019/01/04/debhelper-compat-12-is-now-released/
* https://manpages.ubuntu.com/manpages/jammy/en/man7/debhelper.7.html