Update and change package type to '3.0 (quilt)'
The package type was upgraded from the implicit '1.0 (native)' to the
explicitly defined in debian/source/format '3.0 (quilt)'.

The 3.0 format enables various improvements, including xz compression.

The quilt format is used for regular (non-native) debian packages, and
requires an upstream tarball in the directory above the git repository.
The upstream tarball contains only the source code, which means that
the presence of the .git and the debian directories is not recommended.

For build instructions, please check:
* https://www.debian.org/doc/manuals/debmake-doc/ch05.en.html
* https://wiki.debian.org/Packaging/Intro