Drop unversioned dependencies on essential package
Debian Policy Manual section 3 states that a package is strongly advised
against depending on essential packages without except when explicitly
declaring a version constraint.

Essential packages are defined as "the minimal set of functionality that
must be available and usable on the system at all times" according to
the Debian Policy Manual section 3.8.

Given the reasons mentioned above, the dependencies on coreutils, an
essential package, have been removed from boca-common and
boca-submission-tools. This has also cleared the project from the
Lintian error `depends-on-essential-package-without-using-version`.